Clean up login component
