+++ chat-client/pages/login/view.tsx
@@ -3,19 +3,23 @@ import { Link } from 'react-router-dom';
 import { Row, Col, Card } from 'reactstrap';
 import Form from './form';
 
+interface Props {
+  onSubmit: (email: string, password: string) => void;
+  status: any;
+}
+
 const imageSrc = require('chat-client/shared/images/viking-icon.png');
 
-const LoginView = () => {
-
+const LoginView = ({ onSubmit, status }: Props) => {
   return (
     <div className="d-flex h-100 justify-content-center align-items-center">
       <Card className="p-5">
         <img className="mx-auto" src={imageSrc} height="60" width="55" />
         <h3 className="text-center mt-2 mb-2">Valhalla Chat</h3>
-        <Row style={{ width: 500 }}>
+        <Row>
           <Col>
             <p className="text-center font-weight-bold">Login and start raiding!</p>
-            <Form />
+            <Form onSubmit={onSubmit} status={status} />
           </Col>
         </Row>
         <Row className="justify-content-center font-italic">
